requirements.txt
openai>=1.46.1
pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.8.0
requests>=2.31.0
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.9
